% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-BinaryParticleGenerator.R
\name{BPG}
\alias{BPG}
\title{Binary Particle Generator (BPG)}
\description{
Abstract class for a wrapper making particles for a Binary Swarm
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{chance_bit}}{numeric between 0 and 1
chance to have 1 in the binary construction}

\item{\code{suggestions}}{NULL or list with possible positions}

\item{\code{generator}}{the Object to generate the particles}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BPG-new}{\code{BPG$new()}}
\item \href{#method-BPG-get}{\code{BPG$get()}}
\item \href{#method-BPG-check_suggestions}{\code{BPG$check_suggestions()}}
\item \href{#method-BPG-clone}{\code{BPG$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BPG-new"></a>}}
\if{latex}{\out{\hypertarget{method-BPG-new}{}}}
\subsection{Method \code{new()}}{
Creation of a Binary Particle Generator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BPG$new(generator, chance_bit = 0.5, suggestions = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{generator}}{\code{R6ClassGenerator} for something of class \code{Particle}}

\item{\code{chance_bit}}{numeric between 0 and 1 (not included)
chance to have 1 in the binary construction}

\item{\code{suggestions}}{NULL or list with possible starting positions}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BPG-get"></a>}}
\if{latex}{\out{\hypertarget{method-BPG-get}{}}}
\subsection{Method \code{get()}}{
Creation function for the particles
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BPG$get(size, n_bits, seed = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{integer, number of Particles that needs to be generated}

\item{\code{n_bits}}{integer, number of bits to be represented}

\item{\code{seed}}{NULL or integer, for \code{\link[=set.seed]{set.seed()}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list with selected (\code{suggestions}) and/or generated particles
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BPG-check_suggestions"></a>}}
\if{latex}{\out{\hypertarget{method-BPG-check_suggestions}{}}}
\subsection{Method \code{check_suggestions()}}{
check if all the suggestions can be accepted
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BPG$check_suggestions(suggestions, n_bits)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{suggestions}}{NULL or list with possible starting positions}

\item{\code{n_bits}}{integer, number of bits to be represented}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
logical (\code{logical(1)}) if all suggestions can be accepted
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BPG-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BPG-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BPG$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
