% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-BinaryParticleGenerator.R
\name{BPG_Velocity}
\alias{BPG_Velocity}
\title{Binary Particle Generator-Velocity (BPG-Velocity)}
\description{
Abstract class for a wrapper making particles for a Binary Swarm with Velocity
}
\section{Super class}{
\code{\link[MHFS:BPG]{MHFS::BPG}} -> \code{BPG-Velocity}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{chance_bit}}{numeric between 0 and 1
chance to have 1 in the binary construction}

\item{\code{suggestions}}{NULL or list with possible positions.
Each element of the list should have 2 different elements
First one is for the (binary) position, the second for the velocity
Each position and velocity should have length n_bits, see \code{\link[=get]{get()}}}

\item{\code{boundary_velocity}}{numeric (2 values),
describing the boundaries of the velocity}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BPG-Velocity-new}{\code{BPG_Velocity$new()}}
\item \href{#method-BPG-Velocity-get}{\code{BPG_Velocity$get()}}
\item \href{#method-BPG-Velocity-check_suggestions}{\code{BPG_Velocity$check_suggestions()}}
\item \href{#method-BPG-Velocity-clone}{\code{BPG_Velocity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BPG-Velocity-new"></a>}}
\if{latex}{\out{\hypertarget{method-BPG-Velocity-new}{}}}
\subsection{Method \code{new()}}{
initialize the Binary Particle Generator with Velocity

A subclass of the BPG with Velocity as additional variable whihc needs to be selected and transformed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BPG_Velocity$new(
  generator,
  chance_bit = 0.5,
  boundary_velocity = c(-6, 6),
  suggestions = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{generator}}{\code{R6ClassGenerator} for something of class \code{ParticleVelocity}}

\item{\code{chance_bit}}{numeric between 0 and 1 (not included)
chance to have 1 in the binary construction}

\item{\code{boundary_velocity}}{2 values to where the velocity is bound to.
This is used to generate the velocity randomly between those 2 values}

\item{\code{suggestions}}{NULL or list with possible starting positions}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BPG-Velocity-get"></a>}}
\if{latex}{\out{\hypertarget{method-BPG-Velocity-get}{}}}
\subsection{Method \code{get()}}{
Creation function for the particles
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BPG_Velocity$get(size, n_bits, seed = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{integer, number of Particles that needs to be generated}

\item{\code{n_bits}}{integer, number of bits to be represented}

\item{\code{seed}}{NULL or integer, for \code{\link[=set.seed]{set.seed()}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list with selected (\code{suggestions}) and/or generated particles
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BPG-Velocity-check_suggestions"></a>}}
\if{latex}{\out{\hypertarget{method-BPG-Velocity-check_suggestions}{}}}
\subsection{Method \code{check_suggestions()}}{
check if all the suggestions can be accepted
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BPG_Velocity$check_suggestions(suggestions, n_bits)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{suggestions}}{NULL or list with possible starting positions}

\item{\code{n_bits}}{integer, number of bits to be represented}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
logical (\code{logical(1)}) if all suggestions can be accepted
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BPG-Velocity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BPG-Velocity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BPG_Velocity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
