% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_full_inheritance.R
\name{get_full_inheritance}
\alias{get_full_inheritance}
\title{Get the full inheritance of an object}
\usage{
get_full_inheritance(R6ClassGenerator)
}
\arguments{
\item{R6ClassGenerator}{R6ClassGenerator object}
}
\value{
character with all the names even the start name of the R6ClassGenerator
}
\description{
Get the full inheritance of an object
}
