% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LevyFlight.R
\name{levyFlightProcess}
\alias{levyFlightProcess}
\title{Create Levy flight sequence}
\usage{
levyFlightProcess(n, alpha, gamma = 1, stepsize = 0.01)
}
\arguments{
\item{n}{number of values to be in the constructed}

\item{alpha}{index of the distribution. Special cases are 2=Gaussian, 1=Cauchy, 2/3 and 1/2.}

\item{gamma}{scale unit of the process}

\item{stepsize}{stepsize to take for the levy flight}
}
\value{
sequence of length n
}
\description{
Create Levy flight sequence
}
\references{
Mantegna, Rosario Nunzio.
"Fast, accurate algorithm for numerical simulation of Levy stable stochastic processes."
Physical Review E 49.5 (1994): 4677.
}
