% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHFS-package.R
\docType{package}
\name{MHFS-package}
\alias{MHFS}
\alias{MHFS-package}
\title{MHFS: Meta Heuristic Feature Selection}
\description{
Basic Frame work to implement Population based Meta Heuristic algorithms. Using the OOP principles of the R6 classes.
}
\author{
\strong{Maintainer}: Yucatan De Groote \email{yucatan.de.groote@gmail.com}

}
\keyword{internal}
