% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Particle.R
\name{Particle}
\alias{Particle}
\title{Particle}
\description{
Abstract class for a particle
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{position}}{n-dimensional value which states the position}

\item{\code{track}}{History of positions of this particle}

\item{\code{personal_best}}{Best position or the past}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Particle-new}{\code{Particle$new()}}
\item \href{#method-Particle-save_result}{\code{Particle$save_result()}}
\item \href{#method-Particle-get_personal_best}{\code{Particle$get_personal_best()}}
\item \href{#method-Particle-get_position}{\code{Particle$get_position()}}
\item \href{#method-Particle-set_position}{\code{Particle$set_position()}}
\item \href{#method-Particle-get_track}{\code{Particle$get_track()}}
\item \href{#method-Particle-get_char_position}{\code{Particle$get_char_position()}}
\item \href{#method-Particle-check_position}{\code{Particle$check_position()}}
\item \href{#method-Particle-clone}{\code{Particle$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Particle-new"></a>}}
\if{latex}{\out{\hypertarget{method-Particle-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Particle$new(position)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{position}}{vector with only 0 or 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Particle-save_result"></a>}}
\if{latex}{\out{\hypertarget{method-Particle-save_result}{}}}
\subsection{Method \code{save_result()}}{
save the result of the current position
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Particle$save_result(result)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{result}}{a single numeric value}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Particle-get_personal_best"></a>}}
\if{latex}{\out{\hypertarget{method-Particle-get_personal_best}{}}}
\subsection{Method \code{get_personal_best()}}{
return the best result of the particle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Particle$get_personal_best()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with 2 elements with names
\code{Position} and \code{Result}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Particle-get_position"></a>}}
\if{latex}{\out{\hypertarget{method-Particle-get_position}{}}}
\subsection{Method \code{get_position()}}{
return the current position of the particle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Particle$get_position()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
position, vector of a length
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Particle-set_position"></a>}}
\if{latex}{\out{\hypertarget{method-Particle-set_position}{}}}
\subsection{Method \code{set_position()}}{
return the current position of the particle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Particle$set_position(position)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{position}}{new position of the particle}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
position, vector of a length
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Particle-get_track"></a>}}
\if{latex}{\out{\hypertarget{method-Particle-get_track}{}}}
\subsection{Method \code{get_track()}}{
returns list of the history of this particle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Particle$get_track()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of lists with 2 elements with names
\code{Position} and \code{Result}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Particle-get_char_position"></a>}}
\if{latex}{\out{\hypertarget{method-Particle-get_char_position}{}}}
\subsection{Method \code{get_char_position()}}{
Get string with the current position#'
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Particle$get_char_position()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
string of length \code{length(position)}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Particle-check_position"></a>}}
\if{latex}{\out{\hypertarget{method-Particle-check_position}{}}}
\subsection{Method \code{check_position()}}{
check if the new position only has 0 and 1
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Particle$check_position(position)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{position}}{vector to be checked}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{logical(1)} TRUE if the position only contains 1 and/or 0
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Particle-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Particle-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Particle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
