% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-BAOA.R
\name{ParticleBAOA}
\alias{ParticleBAOA}
\title{R6 Object Representing a Particle for the BAOA algorithm}
\description{
R6 Object Representing a Particle for the BAOA algorithm

R6 Object Representing a Particle for the BAOA algorithm
}
\section{Super class}{
\code{\link[MHFS:Particle]{MHFS::Particle}} -> \code{ParticleBAOA}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ParticleBAOA-clone}{\code{ParticleBAOA$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="MHFS" data-topic="Particle" data-id="check_position"><a href='../../MHFS/html/Particle.html#method-Particle-check_position'><code>MHFS::Particle$check_position()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MHFS" data-topic="Particle" data-id="get_char_position"><a href='../../MHFS/html/Particle.html#method-Particle-get_char_position'><code>MHFS::Particle$get_char_position()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MHFS" data-topic="Particle" data-id="get_personal_best"><a href='../../MHFS/html/Particle.html#method-Particle-get_personal_best'><code>MHFS::Particle$get_personal_best()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MHFS" data-topic="Particle" data-id="get_position"><a href='../../MHFS/html/Particle.html#method-Particle-get_position'><code>MHFS::Particle$get_position()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MHFS" data-topic="Particle" data-id="get_track"><a href='../../MHFS/html/Particle.html#method-Particle-get_track'><code>MHFS::Particle$get_track()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MHFS" data-topic="Particle" data-id="initialize"><a href='../../MHFS/html/Particle.html#method-Particle-initialize'><code>MHFS::Particle$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MHFS" data-topic="Particle" data-id="save_result"><a href='../../MHFS/html/Particle.html#method-Particle-save_result'><code>MHFS::Particle$save_result()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MHFS" data-topic="Particle" data-id="set_position"><a href='../../MHFS/html/Particle.html#method-Particle-set_position'><code>MHFS::Particle$set_position()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParticleBAOA-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ParticleBAOA-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParticleBAOA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
