% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Particle.R
\name{ParticleVelocity}
\alias{ParticleVelocity}
\title{ParticleVelocity}
\description{
Abstract class for a particle with velocity.
This inherits the functionality of a simple particle
}
\section{Super class}{
\code{\link[MHFS:Particle]{MHFS::Particle}} -> \code{ParticleVelocity}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{velocity}}{n-dimensional velocity of the particle}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ParticleVelocity-new}{\code{ParticleVelocity$new()}}
\item \href{#method-ParticleVelocity-get_velocity}{\code{ParticleVelocity$get_velocity()}}
\item \href{#method-ParticleVelocity-set_velocity}{\code{ParticleVelocity$set_velocity()}}
\item \href{#method-ParticleVelocity-set_position_velocity}{\code{ParticleVelocity$set_position_velocity()}}
\item \href{#method-ParticleVelocity-clone}{\code{ParticleVelocity$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="MHFS" data-topic="Particle" data-id="check_position"><a href='../../MHFS/html/Particle.html#method-Particle-check_position'><code>MHFS::Particle$check_position()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MHFS" data-topic="Particle" data-id="get_char_position"><a href='../../MHFS/html/Particle.html#method-Particle-get_char_position'><code>MHFS::Particle$get_char_position()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MHFS" data-topic="Particle" data-id="get_personal_best"><a href='../../MHFS/html/Particle.html#method-Particle-get_personal_best'><code>MHFS::Particle$get_personal_best()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MHFS" data-topic="Particle" data-id="get_position"><a href='../../MHFS/html/Particle.html#method-Particle-get_position'><code>MHFS::Particle$get_position()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MHFS" data-topic="Particle" data-id="get_track"><a href='../../MHFS/html/Particle.html#method-Particle-get_track'><code>MHFS::Particle$get_track()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MHFS" data-topic="Particle" data-id="save_result"><a href='../../MHFS/html/Particle.html#method-Particle-save_result'><code>MHFS::Particle$save_result()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MHFS" data-topic="Particle" data-id="set_position"><a href='../../MHFS/html/Particle.html#method-Particle-set_position'><code>MHFS::Particle$set_position()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParticleVelocity-new"></a>}}
\if{latex}{\out{\hypertarget{method-ParticleVelocity-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Particle with Velocity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParticleVelocity$new(position, velocity)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{position}}{n-dimensional position of the particle}

\item{\code{velocity}}{n-dimensional velocity of the particle}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParticleVelocity-get_velocity"></a>}}
\if{latex}{\out{\hypertarget{method-ParticleVelocity-get_velocity}{}}}
\subsection{Method \code{get_velocity()}}{
Get the current Velocity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParticleVelocity$get_velocity()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
vector of fixed length
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParticleVelocity-set_velocity"></a>}}
\if{latex}{\out{\hypertarget{method-ParticleVelocity-set_velocity}{}}}
\subsection{Method \code{set_velocity()}}{
Set the current velocity into a new one
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParticleVelocity$set_velocity(velocity)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{velocity}}{new velocity of the particle}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParticleVelocity-set_position_velocity"></a>}}
\if{latex}{\out{\hypertarget{method-ParticleVelocity-set_position_velocity}{}}}
\subsection{Method \code{set_position_velocity()}}{
Set the position and velocity into a new one
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParticleVelocity$set_position_velocity(position, velocity)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{position}}{new position of the particle}

\item{\code{velocity}}{new velocity of the particle}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParticleVelocity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ParticleVelocity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParticleVelocity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
