% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_2_dt.R
\name{result_2_dt}
\alias{result_2_dt}
\title{Transform the list of results into a data.table format}
\usage{
result_2_dt(results)
}
\arguments{
\item{results}{list of results based on the \code{run_process()} of the class
\code{BinarySwarm}. Only the element of \code{AllResults} is necessary to give.}
}
\value{
data.table with all the results
}
\description{
Transform the list of results into a data.table format
}
