% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-BAOA.R
\name{SwarmBAOA}
\alias{SwarmBAOA}
\title{R6 Object Representing the Swarm for the BAOA algorithm}
\value{
numeric value

numeric value

matrix with all the new positions.
rows represents a particle.
columns represents the positions in that dimension

NULL.
All positions needs to be changed inside the R6 object
}
\description{
The Swarm based on the Binary Arithmetic Optimizer Algorithm with
}
\section{Super class}{
\code{\link[MHFS:BinarySwarm]{MHFS::BinarySwarm}} -> \code{SwarmBAOA}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{minMoa}}{minimum value for MOA}

\item{\code{maxMoa}}{maximum value for MOA}

\item{\code{beta}}{beta for the BAOA}

\item{\code{k}}{beta for the BAOA}

\item{\code{delta}}{beta for the BAOA}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SwarmBAOA-new}{\code{SwarmBAOA$new()}}
\item \href{#method-SwarmBAOA-MOA}{\code{SwarmBAOA$MOA()}}
\item \href{#method-SwarmBAOA-MOP}{\code{SwarmBAOA$MOP()}}
\item \href{#method-SwarmBAOA-new_positions}{\code{SwarmBAOA$new_positions()}}
\item \href{#method-SwarmBAOA-update_all_positions}{\code{SwarmBAOA$update_all_positions()}}
\item \href{#method-SwarmBAOA-clone}{\code{SwarmBAOA$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="MHFS" data-topic="BinarySwarm" data-id="add_variable_importance"><a href='../../MHFS/html/BinarySwarm.html#method-BinarySwarm-add_variable_importance'><code>MHFS::BinarySwarm$add_variable_importance()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MHFS" data-topic="BinarySwarm" data-id="get_global_best"><a href='../../MHFS/html/BinarySwarm.html#method-BinarySwarm-get_global_best'><code>MHFS::BinarySwarm$get_global_best()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MHFS" data-topic="BinarySwarm" data-id="get_iteration"><a href='../../MHFS/html/BinarySwarm.html#method-BinarySwarm-get_iteration'><code>MHFS::BinarySwarm$get_iteration()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MHFS" data-topic="BinarySwarm" data-id="get_population"><a href='../../MHFS/html/BinarySwarm.html#method-BinarySwarm-get_population'><code>MHFS::BinarySwarm$get_population()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MHFS" data-topic="BinarySwarm" data-id="get_results"><a href='../../MHFS/html/BinarySwarm.html#method-BinarySwarm-get_results'><code>MHFS::BinarySwarm$get_results()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MHFS" data-topic="BinarySwarm" data-id="get_use_variable_importance"><a href='../../MHFS/html/BinarySwarm.html#method-BinarySwarm-get_use_variable_importance'><code>MHFS::BinarySwarm$get_use_variable_importance()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MHFS" data-topic="BinarySwarm" data-id="get_variable_importance"><a href='../../MHFS/html/BinarySwarm.html#method-BinarySwarm-get_variable_importance'><code>MHFS::BinarySwarm$get_variable_importance()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MHFS" data-topic="BinarySwarm" data-id="run_process"><a href='../../MHFS/html/BinarySwarm.html#method-BinarySwarm-run_process'><code>MHFS::BinarySwarm$run_process()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MHFS" data-topic="BinarySwarm" data-id="set_global_best"><a href='../../MHFS/html/BinarySwarm.html#method-BinarySwarm-set_global_best'><code>MHFS::BinarySwarm$set_global_best()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SwarmBAOA-new"></a>}}
\if{latex}{\out{\hypertarget{method-SwarmBAOA-new}{}}}
\subsection{Method \code{new()}}{
initialize the Binary Arithmetic Optimizer Algorithm (BAOA)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SwarmBAOA$new(
  population_size,
  n_bits,
  transferFun,
  particleGenerator,
  beta,
  k,
  minMoa,
  maxMoa,
  delta = 1e-08,
  seed = seed
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{population_size}}{integer, size of the population}

\item{\code{n_bits}}{integer, number of bits to be represented}

\item{\code{transferFun}}{object of class \code{TransferFunction}}

\item{\code{particleGenerator}}{object of class \code{BPG}}

\item{\code{beta}}{Parameter for the BAOA Algoritm}

\item{\code{k}}{Parameter for the BAOA Algoritm}

\item{\code{minMoa}}{Parameter for the BAOA Algoritm}

\item{\code{maxMoa}}{Parameter for the BAOA Algoritm}

\item{\code{delta}}{Parameter for the BAOA Algoritm}

\item{\code{seed}}{seed for the generation of the population}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SwarmBAOA-MOA"></a>}}
\if{latex}{\out{\hypertarget{method-SwarmBAOA-MOA}{}}}
\subsection{Method \code{MOA()}}{
Math optimizer accelerated (MOA)
Function which is used the determine the probability of the exploration or exploitation phase
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SwarmBAOA$MOA(t, Tmax)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{integer, current step of iterations}

\item{\code{Tmax}}{the maximum number of iterations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SwarmBAOA-MOP"></a>}}
\if{latex}{\out{\hypertarget{method-SwarmBAOA-MOP}{}}}
\subsection{Method \code{MOP()}}{
Math optimizer probability (MOP)
Additional value to determine the search
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SwarmBAOA$MOP(t, Tmax)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{integer, current step of iterations}

\item{\code{Tmax}}{the maximum number of iterations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SwarmBAOA-new_positions"></a>}}
\if{latex}{\out{\hypertarget{method-SwarmBAOA-new_positions}{}}}
\subsection{Method \code{new_positions()}}{
Calculate the new positions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SwarmBAOA$new_positions(t, Tmax)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{integer, current step of iterations}

\item{\code{Tmax}}{the maximum number of iterations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SwarmBAOA-update_all_positions"></a>}}
\if{latex}{\out{\hypertarget{method-SwarmBAOA-update_all_positions}{}}}
\subsection{Method \code{update_all_positions()}}{
Update all the positions based on the parameters in the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SwarmBAOA$update_all_positions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SwarmBAOA-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SwarmBAOA-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SwarmBAOA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
