% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-TransferFunctions.R
\name{TransferFunction}
\alias{TransferFunction}
\title{Construction of the S-shaped and V-shaped transfer functions}
\value{
R6 Class object of TransferFunction
}
\description{
Construction of the S-shaped and V-shaped transfer functions

Construction of the S-shaped and V-shaped transfer functions
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character. Name of the transfer function}

\item{\code{fun}}{function. function to transform into values between 0 and 1}

\item{\code{type}}{character, "S" or "V", if it is S-shaped or V-shaped transfer class}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TransferFunction-new}{\code{TransferFunction$new()}}
\item \href{#method-TransferFunction-transfer}{\code{TransferFunction$transfer()}}
\item \href{#method-TransferFunction-changePosition}{\code{TransferFunction$changePosition()}}
\item \href{#method-TransferFunction-get_name}{\code{TransferFunction$get_name()}}
\item \href{#method-TransferFunction-get_type}{\code{TransferFunction$get_type()}}
\item \href{#method-TransferFunction-get_fun}{\code{TransferFunction$get_fun()}}
\item \href{#method-TransferFunction-clone}{\code{TransferFunction$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransferFunction-new"></a>}}
\if{latex}{\out{\hypertarget{method-TransferFunction-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransferFunction$new(name, fun, type = "S")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character. Name of the transfer function}

\item{\code{fun}}{function. function to transform into values between 0 and 1}

\item{\code{type}}{character, "S" or "V", if it is S-shaped or V-shaped transfer class}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
R6 object
Transfer the values
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransferFunction-transfer"></a>}}
\if{latex}{\out{\hypertarget{method-TransferFunction-transfer}{}}}
\subsection{Method \code{transfer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransferFunction$transfer(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{vector or matrix to calculate the function on}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
results of function \code{fun}
Update the next positions based on the transfer functions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransferFunction-changePosition"></a>}}
\if{latex}{\out{\hypertarget{method-TransferFunction-changePosition}{}}}
\subsection{Method \code{changePosition()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransferFunction$changePosition(x, velocity, use_var_importance = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{binary vector of current position}

\item{\code{velocity}}{vector of the next position which is also called the velocity.
based on the velocity the next binary positions are produced}

\item{\code{use_var_importance}}{(Experimental) logical
If variable importance can be done using transfer functions}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
transfer the final results
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransferFunction-get_name"></a>}}
\if{latex}{\out{\hypertarget{method-TransferFunction-get_name}{}}}
\subsection{Method \code{get_name()}}{
Get the name of the Transfer Function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransferFunction$get_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
character with the name
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransferFunction-get_type"></a>}}
\if{latex}{\out{\hypertarget{method-TransferFunction-get_type}{}}}
\subsection{Method \code{get_type()}}{
Get the Type of transfer function (S-shaped or V-shaped)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransferFunction$get_type()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
character with the name
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransferFunction-get_fun"></a>}}
\if{latex}{\out{\hypertarget{method-TransferFunction-get_fun}{}}}
\subsection{Method \code{get_fun()}}{
Get the Transfer Function as function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransferFunction$get_fun()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransferFunction-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TransferFunction-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransferFunction$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
